Create ATA factory to support other modes.
