+++ src/kernel/drivers/storage/ata.cpp
@@ -3,63 +3,19 @@
 #include <type_traits>
 
 #include "drivers/io/ports.hpp"
+#include "drivers/storage/ata_pio.hpp"
 
 namespace drivers::storage {
 
 namespace io = drivers::io;
 
-Ata::Ata(Bus bus, Port port, Mode mode) : port_(port) {}
+Ata::Ata(Bus bus, Port port) : port_(port) {}
 
-void Ata::read_sectors(std::byte* buffer, size_t offset, size_t amount) {
-    send_amount(amount);
-    send_sector_offset(offset);
-    send_command(Command::READ);
-
-    for (size_t sectors_read = 0; sectors_read < amount; sectors_read++) {
-        wait_for_buffer_to_be_ready();
-        read_sector(buffer + (sectors_read * SECTOR_SIZE_IN_BYTES));
-    }
-}
-
-void Ata::send_amount(size_t amount) {
-    io::write_byte(io::Port::PRIMARY_ATA_SELECTOR_COUNT, amount);
-}
-
-void Ata::send_sector_offset(size_t offset) {
-    io::write_byte(io::Port::PRIMARY_ATA_SELECTOR_NUMBER_OR_LBA_LOW, offset);
-    io::write_byte(io::Port::PRIMARY_ATA_CYLINDER_LOW_OR_LBA_MID, offset >> 8);
-    io::write_byte(io::Port::PRIMARY_ATA_CYLINDER_HIGH_OR_LBA_HIGH,
-                   offset >> 16);
-}
-
-void Ata::send_command(Command command) {
-    io::write_byte(io::Port::PRIMARY_ATA_STATUS_OR_COMMAND,
-                   std::underlying_type_t<Command>(command));
-}
-
-void Ata::wait_for_buffer_to_be_ready() {
-    while (!is_buffer_ready()) {
-    }
-}
-
-bool Ata::is_buffer_ready() const {
-    constexpr uint8_t PIO_DATA_AVAILABLE = 0x08;
-
-    const uint8_t status =
-        io::read_byte(io::Port::PRIMARY_ATA_STATUS_OR_COMMAND);
-
-    return status & PIO_DATA_AVAILABLE;
-}
-
-void Ata::read_sector(std::byte* buffer) {
-    constexpr size_t SECTOR_SIZE_IN_WORDS =
-        SECTOR_SIZE_IN_BYTES / (sizeof(uint16_t) / sizeof(uint8_t));
-
-    uint16_t* buffer_as_words = reinterpret_cast<uint16_t*>(buffer);
-
-    for (size_t words_read = 0; words_read < SECTOR_SIZE_IN_WORDS;
-         words_read++) {
-        buffer_as_words[words_read] = io::read_word(io::Port::PRIMARY_ATA_DATA);
+std::unique_ptr<Ata> get_ata(Ata::Mode mode, Ata::Bus bus, Ata::Port port) {
+    switch (mode) {
+        case Ata::Mode::PIO:
+        default:
+            return std::make_unique<AtaPio>(bus, port);
     }
 }
